% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Render a moose twinning report using Quarto}
\usage{
create_report(
  dat_in = paste0(system.file("extdata", package = "moosetwinning"), "/dat.csv"),
  dirname = "quarto"
)
}
\arguments{
\item{dat_in}{a directory path to a CSV file containing moose twinning data (see https://iris.fws.gov/APPS/ServCat/Reference/Profile/179584)}

\item{dirname}{the name of a project directory in which to save the quarto report}
}
\value{
a Word doc
}
\description{
Render a moose twinning report using Quarto
}
\examples{
\dontrun{
create_report(dat_in = paste0(system.file("extdata", package = "moosetwinning"), "/dat.csv"))
}
}
