% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_tbls}
\alias{create_tbls}
\title{Create tables for moose twinning survey reports}
\usage{
create_tbls(dat)
}
\arguments{
\item{dat}{data frame containing moose twinning data, returned by \code{\link[=import_format]{import_format()}}}
}
\value{
a list of flextable objects
}
\description{
Create tables for moose twinning survey reports
}
\examples{
\dontrun{
tbls <- create_tbls(dat)
}
}
