% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{get_report_template}
\alias{get_report_template}
\title{Create a new directory with the moose twinning survey report template}
\usage{
get_report_template(dirname = "quarto")
}
\arguments{
\item{dirname}{character; the name of the directory to create.}
}
\description{
\code{create_quarto_doc} creates a new subdirectory inside the current directory, which will
contain the ready-to-use Quarto file and all associated files. The Word and PDF templates are
based on the standard template of the Alaska Refuge Report Series.
}
\examples{
\dontrun{
 # Create template for Word document
 get_report_template(dirname = "twinning_report")
}
}
