% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_format.R
\name{import_format}
\alias{import_format}
\title{Format Tetlin moose twinning survey for ServCat}
\usage{
import_format(dat_in, boms = FALSE, replace_nas = FALSE)
}
\arguments{
\item{dat_in}{a file directory path to a CSV file containing moose twinning observation data}

\item{boms}{TRUE or FALSE, indicating whether to add required Biological Observation Minimum Specifications fields to the dataframe}
}
\value{
a data frame
}
\description{
Format Tetlin moose twinning survey for ServCat
}
\examples{
\dontrun{
dat <- import_format(dat_in = "./data/dat.csv", boms = T)
}
}
