% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moosetwinning.R
\docType{_PACKAGE}
\name{moosetwinning}
\alias{moosetwinning}
\title{Moose twinning reports}
\description{
An R package that imports moose twinning data, estimates twinning rates, generates plots and summary tables, and produces a survey report. Originally developed for Tetlin National Wildlife Refuge’s moose twinning survey, but could be adopted to other similar surveys. Generates a report that follows the AK Refuge Report Series template.
}
