% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinning_rates.R
\name{twinning_rates}
\alias{twinning_rates}
\title{Calculate moose twinning rates}
\usage{
twinning_rates(dat)
}
\arguments{
\item{dat}{a data frame containing moose twinning data, returned by \code{\link[=import_format]{import_format()}}}
}
\value{
a data frame containing annual and multiyear twinning rates
}
\description{
Calculate moose twinning rates
}
\examples{
\dontrun{
dat_sum <- twinning_rates(dat)
}
}
