% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Write a moose twinning report}
\usage{
create_report(
  dirname = "quarto",
  dat_in = paste0(system.file("extdata", package = "twinning"), "/dat.csv")
)
}
\value{
a Word doc
}
\description{
Write a moose twinning report
}
\examples{
\dontrun{
write_report()
}
}
