% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Render a moose twinning report using Quarto}
\usage{
create_report(
  dat_in = paste0(system.file("extdata", package = "twinning"), "/dat.csv"),
  dirname = "quarto"
)
}
\value{
a Word doc
}
\description{
Render a moose twinning report using Quarto
}
\examples{
\dontrun{
create_report(dat_in = paste0(system.file("extdata", package = "twinning"), "/dat.csv"))
}
}
